minor folders restructure
