+++ dfine_seg/api/coco_names.py
