+++ dfine_seg/app/__init__.py
@@ -1,0 +1,1 @@
+"""The two programs a person launches: the `dfine` command and the Gradio demo."""
