+++ dfine_seg/app/demo.py
@@ -2,7 +2,7 @@
 D-FINE-seg Gradio Demo - detection, instance segmentation, semantic segmentation
 
 Just run it - COCO detection weights download on first use:
-    dfine demo          (or: python -m dfine_seg.demo)
+    dfine demo          (or: python -m dfine_seg.app.demo)
 
 Everything is set from the UI; nothing here needs editing. The "Model" panel swaps in
 your own checkpoint at runtime (size preset or a path/upload) and lets you name its
@@ -35,7 +35,7 @@ import gradio as gr
 import numpy as np
 
 from dfine_seg import load_model
-from dfine_seg.loader import SIZES
+from dfine_seg.api.loader import SIZES
 from dfine_seg.viz import Visualizer, overlay_sem_seg, sem_seg_palette
 
 # ─── Startup defaults (all overridable in the UI) ───────────────────────
